Major overhaul. Prep for package publishing
