+++ README.md
@@ -7,9 +7,9 @@ Groc is a Python CLI designed to help you keep track of purchases. You can enter
 Usage
 --------------
 ```
-./groc [COMMAND] [flags]
+groc [COMMAND] [flags]
 ```
-See also `./groc --help`.
+See also `groc --help`.
 
 
 
@@ -21,7 +21,7 @@ Create a groc database in user's home directory. If a database already exists, t
 
 To see detailed output, use the `--verbose` flag.
 ```
-./groc init
+groc init
 ```
 
 **add**
@@ -37,11 +37,11 @@ To enter purchases via file or directory, use the `--source` flag provided with 
 Adding a purchase that already exists will abort the action, unless the `--ignore-duplicate` flag is passed; this can be especially useful when adding purchases from a file
 or multiple files.
 ```
-./groc add --date 2019-01-01 --total 20.00 --store "Awesome Cakes" --description "birthday cake"
+groc add --date 2019-01-01 --total 20.00 --store "Awesome Cakes" --description "birthday cake"
 
-./groc add --source ./my-purchases/january.csv
+groc add --source ./my-purchases/january.csv
 
-./groc add --source ./my-purchases/ --ignore-duplicate
+groc add --source ./my-purchases/ --ignore-duplicate
 ```
 
 **delete**
@@ -52,7 +52,7 @@ Passing the flag `--dry-run` will output purchases to be deleted without actuall
 
 To see complete purchase details of a purchase, use the `--verbose` flag.
 ```
-./groc delete --id 2 --dry-run
+groc delete --id 2 --dry-run
 ```
 
 **breakdown**
@@ -63,9 +63,9 @@ Target specific months by passing one or multiple month flags like `--month`, `-
 
 To see extended stats, use the `--verbose`.
 ```
-./groc breakdown
+groc breakdown
 
-./groc breakdown --month=01 --month=03 --year=2019
+groc breakdown --month=01 --month=03 --year=2019
 ```
 
 **list**
@@ -77,9 +77,9 @@ To see all purchases of a month, pass the `--all`, `-a` flag.
 
 To see detailed output, such as purchase id, use the `--verbose` flag.
 ```
-./groc list --limit 10
+groc list --limit 10
 
-./groc list -m 02 --all
+groc list -m 02 --all
 ```
 
 **reset**
@@ -88,5 +88,5 @@ Reset a groc database by deleting all entries. The database and schema will not 
 
 Passing the `--dry-run` flag will output the purchase count to be reset.
 ```
-./groc reset --verbose
+groc reset --verbose
 ```
