--- groc
@@ -1,3 +1,0 @@
-#!/usr/bin/env bash
-
-python -m app.main "$@"
