+++ groc/__main__.py
@@ -1,0 +1,6 @@
+from .cli import safe_entry_point
+
+
+"""groc.__main__: executed when groc directory is called as script."""
+
+safe_entry_point()
